LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/min-inline-size-expected.txt





PASS horizontal-tb
FAIL vertical-lr  assert_equals: height expected "100px" but got "0px"
FAIL vertical-rl  assert_equals: height expected "100px" but got "0px"
PASS horizontal-tb override
PASS vertical-lr override
PASS vertical-rl override
